refactor : 패키지 이름 변경 common -> console
